# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Citas Médicas - DLYTime
  description: API REST para la gestión de usuarios, autenticación, citas, historial clínico, empleados y clientes.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Servidor local de desarrollo
...
paths:
...
  /crudCitas/crudCita:
    get:
      tags:
        - CrudCita
      summary: Obtener citas filtradas
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: fecha
          schema:
            type: string
        - in: query
          name: idTipoConsulta
          schema:
            type: string
      responses:
        '200':
          description: Lista de citas obtenida exitosamente
  /crudCitas/addCita:
    post:
      tags:
        - CrudCita
      summary: Agregar una nueva cita
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                NumeroDocumentoCliente:
                  type: string
                NumeroDocumentoOftalmologo:
                  type: string
                idHorario:
                  type: integer
                idTipoConsulta:
                  type: integer
      responses:
        '201':
          description: Cita agregada exitosamente
        '500':
          description: Error durante la transacción
  /crudCitas/updateCita:
    patch:
      tags:
        - CrudCita
      summary: Actualizar una cita existente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idCita
                - idHorario
                - idTipoConsulta
                - idHorario1
              properties:
                idCita:
                  type: integer
                idHorario:
                  type: integer
                idTipoConsulta:
                  type: integer
                idHorario1:
                  type: integer
      responses:
        '201':
          description: Cita actualizada exitosamente
        '500':
          description: Error durante la transacción
  /crudCitas/cancelCita:
    patch:
      tags:
        - CrudCita
      summary: Cancelar una cita
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idCita
                - idEstadoCita
                - idHorario
                - idEstadoHorario
              properties:
                idCita:
                  type: integer
                idEstadoCita:
                  type: integer
                idHorario:
                  type: integer
                idEstadoHorario:
                  type: integer
      responses:
        '201':
          description: Cita cancelada
        '400':
          description: No se puede activar una cita ya atendida
        '500':
          description: Error durante la transacción
...